Write out to Dolma
